zsmalloc: add to mm/
=========
DO NOT MERGE, FOR REVIEW ONLY
This patch introduces zsmalloc as new code, however, it already
exists in drivers/staging.  In order to build successfully, you
must select EITHER to driver/staging version OR this version.
Once zsmalloc is reviewed in this format (and hopefully accepted),
I will create a new patchset that properly promotes zsmalloc from
staging.
=========

This patchset introduces a new slab-based memory allocator,
zsmalloc, for storing compressed pages.  It is designed for
low fragmentation and high allocation success rate on
large object, but <= PAGE_SIZE allocations.

zsmalloc differs from the kernel slab allocator in two primary
ways to achieve these design goals.

zsmalloc never requires high order page allocations to back
slabs, or "size classes" in zsmalloc terms. Instead it allows
multiple single-order pages to be stitched together into a
"zspage" which backs the slab.  This allows for higher allocation
success rate under memory pressure.

Also, zsmalloc allows objects to span page boundaries within the
zspage.  This allows for lower fragmentation than could be had
with the kernel slab allocator for objects between PAGE_SIZE/2
and PAGE_SIZE.  With the kernel slab allocator, if a page compresses
to 60% of it original size, the memory savings gained through
compression is lost in fragmentation because another object of
the same size can't be stored in the leftover space.

This ability to span pages results in zsmalloc allocations not being
directly addressable by the user.  The user is given an
non-dereferencable handle in response to an allocation request.
That handle must be mapped, using zs_map_object(), which returns
a pointer to the mapped region that can be used.  The mapping is
necessary since the object data may reside in two different
noncontigious pages.

zsmalloc fulfills the allocation needs for zram and zswap.

Acked-by: Nitin Gupta <ngupta@vflare.org>
Acked-by: Minchan Kim <minchan@kernel.org>
Signed-off-by: Seth Jennings <sjenning@linux.vnet.ibm.com>