From	Matthew Garrett <> Date	Sat, 18 Apr 2015 08:26:36 -0700 Subject	[PATCH 3/3] libata: Change medium_power LPM policy to match Intel recommendations
Intel publish a document on designing energy efficient SATA devices at
http://www.intel.com/content/dam/doc/reference-guide/sata-devices-implementation-recommendations.pdf
which recommends that ALPE be set, ASPE be cleared and that DIPM be
enabled
on the device. Right now we have no policy that matches that -
medium_power
does not enable DIPM and min_power sets ASPE. Change medium_power to
implement these recommendations, with the addition of devslp state being
inherited from the initial configuration. With luck this will provide
reasonable power savings without causing the device breakages we
occasionally see with the min_power policy.