muqss: Tune CPU selection v2
I was testing MuQSS for performance using MC, LLC and SMT and discovered that
it is not up to performance, at least on my machine. This commit fixes the
performance, at least on my machine (Ryzen 1700), I have no way of checking
that on any Intel machine as I don't have one I can test on, but it should
work similarly.

This patch modifies best CPU selection:
    Adds a check whether not only the CPU in question is in the
    idle mask, but it will check whether the mask contains its
    SMT sibling too.
    The rationale is that SMT shares all resources of the core
    and if other task is already scheduled there, it's not
    really optimal to schedule more tasks that share the same
    SMT resources.

    Patch also refines further best CPU selection and adds an
    exit threshold.
    The rationale there is that there is no need to search
    further if the best selection is already found. The best
    selection depends on localities processed and is mainly
    targeted around SMT.