mm: fix priority queue in multigenerational lru
mm: fix priority queue in multigenerational lru                                                         
                                                                                                        
The priority queue helps select the best memcg to reclaim memory from                                   
when there are multiple candidates. If reclaim didn't make progress                                     
after it has tried those at the front, it increments the priority of                                    
the rest in the queue by one. If the queue is not empty, more memcgs                                    
will appear at the front and then reclaim will try them again.                                          
                                                                                                        
However, this wasn't properly implemented, which causes the memory                                      
pressure to be distributed to all memcgs in the queue. If one of them                                   
uses a large amount of page cache, the distributed memory pressure                                      
will slow down the rest.                                                                                
                                                                                                        
Signed-off-by: Yu Zhao <yuzhao@google.com>