ZEN: sched/alt: Never yield unless 2 or more tasks are running
Looking over the yield code in CFS, the way mainline handles yielding is
it just assumes there's no reason to reschedule if only one task is
running.

Although that condition is checked in yield type 2, for some reason it's
missing from yield type 1 as well.  Considering there's zero benefit to
rescheduling when there's no contention on the run-queue, always do
nothing when 1 or less tasks are running on the RQ.