ZEN: sched/alt: Further optimize `do_sched_yield`
Invert nr_running condition so we encapsulate all conditional code
inside if statement.  This also lets us move the `schedstat_inc(...)`
invocation inside and conditionally execute if we actually call yield.

Also, remove the test on `!rt_task(...)`.  Although this may be valid,
it requires we add a boolean to know if we truly yield since there's a
chance we won't yield inside the `if(rq->nr_running > 1)` block.  Not to
mention, what's the chance a realtime task wants to yield?  It's not
zero but probably won't be an issue.

Thanks to @torvic9 for the idea:
 - https://gitlab.com/alfredchen/linux-prjc/-/issues/87#note_1670659217