sched/alt: Reintroduce CONFIG_FAIR_GROUP_SCHED stubs and sync-up
Although Project-C doesn't support or implement anything part of
CONFIG_FAIR_GROUP_SCHED, the interfaces are still needed for software
like Kubernetes.

Reintroduce the interfaces as stubs so we don't break compatibility.

Also, sync-up #ifdef macros with upstream so that stubs that are not
necessary are not compiled in if the feature stub is not wanted.